+++ android/app-tv/src/main/kotlin/app/projektor/tv/ui/player/TvPlayerScreen.kt
@@ -77,7 +77,7 @@ fun TvPlayerScreen(container: TvContainer, fileId: String, itemId: String, start
     var detail by remember { mutableStateOf<ItemDetail?>(null) }
     var file by remember { mutableStateOf<MediaFile?>(null) }
     var decision by remember { mutableStateOf<PlaybackDecision?>(null) }
-    var audioIndex by remember { mutableStateOf<Int?>(null) }
+    var audioIndex by remember { mutableStateOf<Long?>(null) }
     var error by remember { mutableStateOf<String?>(null) }
     var controlsVisible by remember { mutableStateOf(false) }
     var lastInteraction by remember { mutableStateOf(System.currentTimeMillis()) }
@@ -109,7 +109,7 @@ fun TvPlayerScreen(container: TvContainer, fileId: String, itemId: String, start
     LaunchedEffect(file, audioIndex) {
         val f = file ?: return@LaunchedEffect
         try {
-            val d = items.decide(PlaybackDecideRequestInput(fileId = f.id, profile = DeviceProfiles.current(), audioStreamIndex = audioIndex, startPositionMs = resumeAt.toInt()))
+            val d = items.decide(PlaybackDecideRequestInput(fileId = f.id, profile = DeviceProfiles.current(), audioStreamIndex = audioIndex, startPositionMs = resumeAt))
             decision = d
             player.load(mediaSpecFor(d, client), startMs = resumeAt, rate = settings.rate, knownDurationMs = f.durationMs.toLong())
             rootFocus.requestFocus()
